MedicinesActivity: don't wait for layout inflation before starting search
This was done as a workaround to avoid null pointers when
onActivityResult was called. A side effect of this workaround is that
there was sometimes a flash of MedicinesActivity just before the search
appeared.

Instead, let's wait for the layout at onActivityResult. It avoids both
the NPEs and the ugly flash.