RoutinesActivityCreate/Edit test: Solve autocorrect bug
In certain phones, the input will auto-capitalize the first letter of
words. In this case, the test will fail as the routine is found by name,
and it will not match the input.

These tests are meant to find out if the routines are correctly saved,
not if the text input is what we expect, so we'll just pre-capitalize
these strings, and it should work anywhere.